removed bot.js and consolidate in app.js
